+++ CHANGELOG.md
@@ -18,6 +18,12 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Fixed custom base URL-s with providers
 * Fixed `messageHelpers` type
 * Fixed `State.initFromRouter`
+* [Fixed `reInit` method in `b-virtual-scroll`](https://github.com/V4Fire/Client/pull/162)
+* [Fixed `in-view` initialization](https://github.com/V4Fire/Client/pull/162)
+
+#### :house: Internal
+
+* [Added `unsafe` getter for `b-virtual-scroll`, `b-virtual-scroll/scroll-render`, `b-virtual-scroll/scroll-request`](https://github.com/V4Fire/Client/pull/162)
 
 ## v3.0.0-beta.250 (2020-02-12)
 
@@ -48,7 +54,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :rocket: New Feature
 
-* [Added `i-item` trait](https://github.com/V4Fire/Client/pull/154)
+* [Added `i-items` trait](https://github.com/V4Fire/Client/pull/154)
 * [Added `runtime` block to `b-skeleton`](https://github.com/V4Fire/Client/pull/154)
 * [Added `inViewFactory` to create local `in-view` instances](https://github.com/V4Fire/Client/pull/154)
 * [Added `observable.size` for `in-view` observable elements](https://github.com/V4Fire/Client/pull/154)